Add clang-tidy support (#97)
Added clang-tidy checks locally and via CI, resolved all code that made sense to. Ignored the bridge.

+++ include/livekit/data_track_stream.h
@@ -124,7 +124,7 @@ private:
   FfiHandle subscription_handle_;
 
   /** FfiClient listener id for routing FfiEvent callbacks to this object. */
-  std::int64_t listener_id_{0};
+  std::int32_t listener_id_{0};
 };
 
 } // namespace livekit
